Created separate file for mouse tracker.
